refactor: 파일 읽기 추상화
